requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
tree-sitter>=0.22.0
tree-sitter-javascript>=0.23.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
